WIP: parsing args, now segfaulting
